modifications

+++ ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: estina <estina@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 12:37:07 by estina            #+#    #+#             */
-/*   Updated: 2019/11/04 12:41:24 by estina           ###   ########.fr       */
+/*   Updated: 2019/11/04 20:26:20 by estina           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
 	if (n == 0)
 		return ;
