+++ ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: estina <estina@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 12:42:19 by estina            #+#    #+#             */
-/*   Updated: 2019/11/04 12:54:02 by estina           ###   ########.fr       */
+/*   Updated: 2019/11/04 20:32:04 by estina           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, int n)
+#include "libft.h"
+
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	while (n > 0)
 	{
