+++ ft_memmove.c
@@ -6,13 +6,15 @@
 /*   By: estina <estina@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 13:11:39 by estina            #+#    #+#             */
-/*   Updated: 2019/11/04 14:35:33 by estina           ###   ########.fr       */
+/*   Updated: 2019/11/04 20:50:39 by estina           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dst, const void *src, int len)
+#include "libft.h"
+
+void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if (src == dst)
