+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: estina <estina@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 18:21:42 by estina            #+#    #+#             */
-/*   Updated: 2019/11/04 19:17:07 by estina           ###   ########.fr       */
+/*   Updated: 2019/11/04 20:51:01 by estina           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strncmp(const char *s1, const char *s2, int n)
+#include "libft.h"
+
+int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n && (s1[i] || s2[i]))
